% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{compute_probabilities}
\alias{compute_probabilities}
\title{Compute the probabilites after EM}
\usage{
compute_probabilities(data, means, variances, proportions)
}
\arguments{
\item{data}{\code{n} by \code{p} matrix, where \code{n} is the number of
samples and \code{p} is the number of features}

\item{means}{a matrix with \code{p} columns and \code{k} rows (for the
number of clusters), where each row represents the mean of a Spherical
Gaussian}

\item{variances}{a non-negative vector of length \code{k}, where each value
denotes the variance of the Spherical Gaussian}

\item{proportions}{a non-negative vector of length \code{k} that sums to 1,
where each value denotes the relative proportion of each cluster}
}
\value{
a matrix of \code{n} by \code{k}, where each row denotes the
probability that each sample was drawn from that cluster. Each row sums to 1
and is non-negative
}
\description{
Specifically, compute the probabilities of each sample originating from the
Spherical Gaussian (after running \code{EM_spherical_gaussians})
}
\examples{
n <- 500
means <- matrix(c(0, 0, 5, 5), ncol = 2, byrow = TRUE)
covariances <- list(diag(2),
                    diag(2))
proportions <- c(0.5, 0.5)
data <- generate_mixture_gaussians(n, means, covariances, proportions)

# Fit EM algorithm to the generated data
k <- 2 # Number of Gaussian components
em_results <- EM_spherical_gaussians(data, k)
means_estimated <- em_results$means
variances_estimated <- em_results$variances
proportions_estimated <- em_results$proportions

# Compute the probabilities for each sample
probabilities <- compute_probabilities(data = data,
                                       means = means_estimated,
                                       variances = variances_estimated,
                                       proportions = proportions_estimated)
head(probabilities)
}
