% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_square_func.R
\name{custom_square_func}
\alias{custom_square_func}
\title{Squared function}
\usage{
custom_square_func(x)
}
\arguments{
\item{x}{scalar value}
}
\value{
a scalar value
}
\description{
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
v
This is my demo squared function. 
This is my demo squared function. 
This is my demo squared function. 
v
v
This is my demo squared function.
}
\examples{
custom_square_func(10)
}
