% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{EM_spherical_gaussians}
\alias{EM_spherical_gaussians}
\title{EM algorithm for Spherical Gaussians}
\usage{
EM_spherical_gaussians(data, k, max.iter = 100, tol = 1e-06)
}
\arguments{
\item{data}{\code{n} by \code{p} matrix, where \code{n} is the number of
samples and \code{p} is the number of features}

\item{k}{number of clusters}

\item{max.iter}{maximum iteration of EM}

\item{tol}{relative change in log-likelihood. If the change is less than
\code{tol}, then the method terminates early.}
}
\value{
a list with elements \code{means}, \code{variances},
\code{proportions}, and \code{log_likelihood} for all the respective
estimated components of the mixture of Spherical Gaussians
}
\description{
EM algorithm for Spherical Gaussians
}
\examples{
n <- 500
means <- matrix(c(0, 0, 5, 5), ncol = 2, byrow = TRUE)
covariances <- list(diag(2),
                    diag(2))
proportions <- c(0.5, 0.5)
data <- generate_mixture_gaussians(n, means, covariances, proportions)

# Fit EM algorithm to the generated data
k <- 2 # Number of Gaussian components
em_results <- EM_spherical_gaussians(data, k)
em_results
}
