% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_mixture_gaussians}
\alias{generate_mixture_gaussians}
\title{Simulate a mixture of Spherical Gaussians}
\usage{
generate_mixture_gaussians(n, means, covariances, proportions)
}
\arguments{
\item{n}{number of samples}

\item{means}{a matrix with \code{p} columns and \code{k} rows (for the
number of clusters), where each row represents the mean of a Spherical
Gaussian}

\item{covariances}{a list of \code{k} \code{p} positive definite matrices}

\item{proportions}{a non-negative vector of length \code{k} that sums to 1,
where each value denotes the relative proportion of each cluster}
}
\value{
a \code{n} by \code{p} matrix, where \code{n} is the number of
samples and \code{p} is the number of features
}
\description{
Simulate a mixture of Spherical Gaussians
}
\examples{
n <- 500
means <- matrix(c(0, 0, 5, 5), ncol = 2, byrow = TRUE)
covariances <- list(diag(2),
                    diag(2))
proportions <- c(0.5, 0.5)
data <- generate_mixture_gaussians(n, means, covariances, proportions)
head(data)
}
